+++ .github/scripts/install-deno.sh
@@ -3,7 +3,7 @@
 # Installs a binary distribution of Deno. (The Snapcraft version of Deno seems
 # unable to locate `cargo`.)
 
-version=1.17.1
+version=1.22.3
 wget https://github.com/denoland/deno/releases/download/v${version}/deno-x86_64-unknown-linux-gnu.zip
 unzip deno-x86_64-unknown-linux-gnu.zip
 chmod +x deno