+++ src/r3_port_std/src/threading_unix.rs
@@ -33,8 +33,7 @@ pub unsafe fn exit_thread() -> ! {
 /// [`std::thread::JoinHandle`] with extra functionalities.
 #[derive(Debug)]
 pub struct JoinHandle<T> {
-    #[allow(dead_code)]
-    std_handle: thread::JoinHandle<T>,
+    _std_handle: thread::JoinHandle<T>,
     thread: Thread,
 }
 
@@ -69,7 +68,10 @@ pub fn spawn(f: impl FnOnce() + Send + 'static) -> JoinHandle<()> {
     // Wait until the just-spawned thread configures its own `THREAD_DATA`.
     thread::park();
 
-    JoinHandle { std_handle, thread }
+    JoinHandle {
+        _std_handle: std_handle,
+        thread,
+    }
 }
 
 impl<T> JoinHandle<T> {