+++ src/r3_support_rp2040/src/usbstdio.rs
@@ -79,8 +79,8 @@ pub const fn configure<'pool, C, TOptions: Options>(
     rp2040_resets: Bind<'pool, C::System, rp2040_pac::RESETS>,
     rp2040_usbctrl_regs: Bind<'pool, C::System, rp2040_pac::USBCTRL_REGS>,
 ) where
-    C: ~const traits::CfgBase + ~const traits::CfgInterruptLine,
-    C::System: traits::KernelInterruptLine + traits::KernelStatic,
+    // `~const CfgBase` not implied due to [ref:const_supertraits]
+    C: ~const traits::CfgBase + ~const traits::CfgStatic + ~const traits::CfgInterruptLine,
 {
     bind(
         (rp2040_resets.borrow_mut(), rp2040_usbctrl_regs.take()),